Correctly detect MetaData Support
With the current version, MetaData support is not detected in (for example) AirSonos
This is because the 'icy-metadata' header appears to be a String value of "1", rather than a numeric value of 1.
Reverting the comparison back to "==" instead of "===" relaxes the comparison, and allows AirSonos to send the song title etc. to the Sonos.